refactor(connectors): own the resolution machinery, inject the connector map
The connector registry, config parsing and the local password DB moved into the
connectors package as connectors.Resolver(storage, logger, configs), which
returns the ResolveFunc the cache already takes. The config map is injected by
the caller, so the connectors package imports no connector implementation and a
library consumer can supply its own set. server keeps only ConnectorsConfig (the
built-in map) plus ConnectorConfig/LocalConnector aliases for compatibility.
NewPasswordDB is exported so a custom resolver can reuse the local connector.
